[hardware] Restrict mem bus to EW if vstore, vstart > 0, and EW < 64-bit
If vstart > 0 and EW < 64, the situation is similar to when the memory addr
is misaligned wrt the memory bus. Because of the VRF Byte Layout and
since the granularity of each lane's payload to the store unit is 64 bit,
all the packets can contain valid data while we have not completed the
beat. So, either we calculate in the addrgen the effective length of a
bursts with unequal beats, or we add a buffer and aligner in the store
unit, or we handle the ready signals at a byte level, or we simply reduce
the effective memory bus to the element width (worst case).
We do the latter. It's low performance, but vstore with vstart > 0 happen
after an exception, so the throughput drop should be acceptable.