[hardware] :bug: Don't use vstart to drop elements for slides
The vstart signal within the lanes is not the architectural
vstart. For all the instructions, it corresponds to
the architectural vstart manipulated to reflect the "vstart"
in every lane for VRF fetch address calculation purposes.
Memory instructions, which support arch vstart > 0, can use
that vstart signal to resize the number of elements to fetch
from the VRF. Slide instructions, instead, further modify
the vstart only for addressing purposes, and should not use
the vstart signal to resize the number of elements to fetch.